+++ spec/rubocop/cop/lint/format_parameter_mismatch_spec.rb
@@ -200,6 +200,44 @@ describe RuboCop::Cop::Lint::FormatParameterMismatch do
     expect(cop.offenses).to be_empty
   end
 
+  context 'on format with %{} interpolations' do
+    context 'and 1 argument' do
+      it 'does not register an offense' do
+        inspect_source(cop, ["params = { y: '2015', m: '01', d: '01' }",
+                             "puts format('%{y}-%{m}-%{d}', params)"])
+        expect(cop.offenses).to be_empty
+      end
+    end
+
+    context 'and multiple arguments' do
+      it 'registers an offense' do
+        inspect_source(cop, ["params = { y: '2015', m: '01', d: '01' }",
+                             "puts format('%{y}-%{m}-%{d}', 2015, 1, 1)"])
+        expect(cop.messages).to eq(['Number of arguments (3) to `format` ' \
+                                    "doesn't match the number of fields (1)."])
+      end
+    end
+  end
+
+  context 'on format with %<> interpolations' do
+    context 'and 1 argument' do
+      it 'does not register an offense' do
+        inspect_source(cop, ["params = { y: '2015', m: '01', d: '01' }",
+                             "puts format('%<y>d-%<m>d-%<d>d', params)"])
+        expect(cop.offenses).to be_empty
+      end
+    end
+
+    context 'and multiple arguments' do
+      it 'registers an offense' do
+        inspect_source(cop, ["params = { y: '2015', m: '01', d: '01' }",
+                             "puts format('%<y>d-%<m>d-%<d>d', 2015, 1, 1)"])
+        expect(cop.messages).to eq(['Number of arguments (3) to `format` ' \
+                                    "doesn't match the number of fields (1)."])
+      end
+    end
+  end
+
   it 'finds the correct number of fields' do
     expect(''.scan(described_class::FIELD_REGEX).size)
       .to eq(0)